requirements.txt
discord-py==2.6.4
python-dotenv==1.1.1
openai==2.6.0